lib/libutil: add overflow checking for memory allocation functions
The "normal" *alloc() functions are commonly used in a pattern like this:

  malloc(sizeof(int) * var);

which is pretty dangerous if you start considering that such a calculation
might overflow. Quite a few codebases added the reallocarray(3) function to
mitigate this issue. I decided to make it mandatory to pass two parameters
to _every_ memory allocation function, as I'm convinced that this API is
better. As a bonus, it's harder for people not knowing about this problem
to make mistakes in this regard.