remove LinuxStandardBase
lsb is a standard too irrelevant to dedicate a complete subdirectory to it.
Most utilities described there are either useful for non-linux systems too
or are better placed into a 'linux' subdirectory that contains way more
linux specific tools (e.g. merged with a selection of util-linux clones)
than just the lsb ones.