const-qualify pointers and attempt to use const-correctness in general
I changed my mind on const in C. I now think it adds value and improves
communicating the intent of the programmer. Additionally, it allows the
compiler to provide more warnings and errors.
As for the syntax I decided to use the following

  char const *p;

because it can be read from right to left (pointer [to] constant char).