libutil: add a non-blocking fread(3) alternative
Most utilities rely on non-blocking reads but stdio doesn't support them.
The tools _could_ call fread(3) with an nmemb argument of 1 to avoid
waiting for input that might never arrive, but that's inefficient and somewhat
ugly. Thus, wrapping read(2) in a stdio-like interface seems like a good
solution because it hides the details in libutil while the utility code stays
clean.